Separate check for updates button and update notification
